+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add rate limiting for multiple messages with `--rate` flag
+
 ## 1.3.0 - 2019-11-13
 
 ### Added